feat: implement safe API functions for Chromium AI and enhance TypeScript configuration
