+++ tsconfig.json
@@ -8,6 +8,17 @@
 		"skipLibCheck": true,
 		"forceConsistentCasingInFileNames": true,
 		"declaration": true,
+		"noUnusedLocals": true,
+		"noUnusedParameters": true,
+		"noImplicitReturns": true,
+		"noFallthroughCasesInSwitch": true,
+		"noUncheckedIndexedAccess": true,
+		"noImplicitOverride": true,
+		"allowUnusedLabels": false,
+		"allowUnreachableCode": false,
+		"isolatedModules": true,
+		"useDefineForClassFields": true,
+		"erasableSyntaxOnly": true,
 		"outDir": "./dist",
 		"lib": ["ESNext", "DOM"]
 	},
